msvc: Fix build errors with /sdl compiler switch
Being an internet facing application putty and company should be built
with /sdl for the additional compiler checks as they stand to benefit
from them.  This includes enabling strict_gs_check and freed pointer
sanitation in addition to the warnings/errors.  This is enabled for
Visual Studio 2010/2012 generated project files.

One method does not appear to be fixable possibly due to a compiler bug.
The appropriate warning is disabled for that method.

Signed-off-by: Alex Smith <alex.smith@warpsharp.info>